requirements.txt
google-auth==2.31.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-generativeai==0.3.2
pyppeteer==1.0.2